Fixes and clean-up: currently passing all tile tests and all mesh tests, tested on a 2x2 mesh
Update README.md

Delete doc/MAGIA.png

Add files via upload

Update README.md

Update README.md

+++ sw/tests/l1_test.c
@@ -13,7 +13,7 @@
 #define ACTIVE_WORDS ((L1_SIZE/(BITS_WORD/BITS_BYTE))/NUM_L1_BANKS)
 
 int main(void) {
-  uint8_t exit_code;
+  uint32_t exit_code;
 
 #ifdef FULL_L1
   uint32_t expected_result = (NUM_L1_BANKS*ACTIVE_WORDS/STEP)*(NUM_L1_BANKS*ACTIVE_WORDS/STEP + 1)/2;
